Tests/LibWeb/Text/input/UIEvents/pointer-boundary-events.html

<!DOCTYPE html>
<style>
body {
    margin: 0;
    padding: 5px;
    width: 200px;
    height: 200px;
}

#outer {
    width: 100px;
    height: 100px;
    background-color: yellowgreen;
}

#inner {
    width: 50px;
    height: 50px;
    background-color: magenta;
}
</style>
<body id="body"><div id="outer"><div id="inner"></div></div></body>
<script src="../include.js"></script>
<script>
    function logEvent(e) {
        println(`${e.type} target.id=(${e.target.id}) relatedTarget.id=(${e.relatedTarget?.id ?? 'null'})`);
    }

    const events = [
        'pointerover',
        'pointerout',
        'pointerenter',
        'pointerleave',
        'pointermove',
    ];

    events.forEach(eventType => {
        outer.addEventListener(eventType, logEvent);
        inner.addEventListener(eventType, logEvent);
    });

    test(() => {
        internals.movePointerTo(150, 150);

        println("> move pointer over #inner");
        internals.movePointerTo(10, 10);

        println("> move pointer over #outer (leaving #inner)");
        internals.movePointerTo(60, 60);

        println("> move pointer outside (leaving #outer)");
        internals.movePointerTo(150, 150);
    });
</script>
