Tests/LibWeb/Text/expected/UIEvents/pointer-boundary-events.txt

> move pointer over #inner
pointerover target.id=(inner) relatedTarget.id=(body)
pointerover target.id=(inner) relatedTarget.id=(body)
pointerenter target.id=(outer) relatedTarget.id=(body)
pointerenter target.id=(inner) relatedTarget.id=(body)
pointermove target.id=(inner) relatedTarget.id=(null)
pointermove target.id=(inner) relatedTarget.id=(null)
> move pointer over #outer (leaving #inner)
pointerout target.id=(inner) relatedTarget.id=(outer)
pointerout target.id=(inner) relatedTarget.id=(outer)
pointerleave target.id=(inner) relatedTarget.id=(outer)
pointerover target.id=(outer) relatedTarget.id=(inner)
pointermove target.id=(outer) relatedTarget.id=(null)
> move pointer outside (leaving #outer)
pointerout target.id=(outer) relatedTarget.id=(out)
pointerleave target.id=(outer) relatedTarget.id=(out)
